Disassociate via the actors so they shutdown correctly
When a local connection is dropped we now shutdown the local actor
and so unsubscribe from the connection.  There is still a race
condition the subscription addresses are not unique so a listener
failure will likely try to re-create the listener before the old
one has completed resulting in an actor name collison.